hw_lesson 24/online_shop/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.4.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
